--- 5-1.py
@@ -1,34 +1,0 @@
-# Пользователь вводит данные о количестве предприятий, их наименования и прибыль за 4 квартала (т.е. 4 отдельных числа)
-# для каждого предприятия.. Программа должна определить среднюю прибыль (за год для всех предприятий)
-# и вывести наименования предприятий, чья прибыль выше среднего и отдельно вывести наименования предприятий,
-# чья прибыль ниже среднего.
-
-business_all = int(input("Пожалуйста, введите натуральное число, отображающее колличество предприятий "))
-income_all = 0
-bincome = {}
-
-for i in range(business_all):
-    business_name = input(f"Пожалуйста, введите название {i+1} предприятия ")
-    income_1 = int(input("Пожалуйста, введите целое число -  прибыль за первый квартал "))
-    income_2 = int(input("Пожалуйста, введите целое число -  прибыль за второй квартал "))
-    income_3 = int(input("Пожалуйста, введите целое число -  прибыль за третий квартал "))
-    income_4 = int(input("Пожалуйста, введите целое число -  прибыль за четвертый квартал "))
-    bincome[business_name] = income_1 + income_2 + income_3 + income_4
-    income_all += income_1 + income_2 + income_3 + income_4
-
-avg_income = income_all / business_all
-print(f"Средняя прибыль по всем предприятиям за год равна {avg_income}")
-
-for i in bincome:
-    if bincome[i] < avg_income:
-        print(f"Прибыль предприятия {i} ниже среднего")
-    elif bincome[i] == avg_income:
-        print(f"Прибыль предприятия {i} средняя")
-    else:
-        print(f"Прибыль предприятия {i} выше среднего")
-
-
-
-
-
-
