+++ Homework_1/ДЗ к 1 уроку/1-1.py
