+++ Homework_1/ДЗ к 1 уроку/1-3.py
