+++ Homework_1/ДЗ к 1 уроку/1-4.py
