+++ Homework_2/2-1.py
