+++ Homework_2/2-2.py
