+++ Homework_2/2-3.py
