+++ Homework_2/2-4.py
