+++ Homework_2/2-5.py
