+++ Homework_2/2-6.py
