+++ Homework_2/Task_2.txt
