+++ Homework_3/3-3.py
