+++ Homework_3/3-5.py
