+++ Homework_3/3-6.py
